+++ UnitTests/TestUtils.h
@@ -33,7 +33,7 @@ public:
 
    virtual void OpenFile(const char* config_file);
    virtual void SetConfiguration(const char* section, const char* cle, const char* valeur, const char* file);
-   virtual unsigned int GetConfiguration(const char* section, const char* cle, const char* default_value, char* out_buffer, size_t buffer_size, const char* file);
+   virtual unsigned int GetConfiguration(const char* section, const char* cle, const char* default_value, char* out_buffer, unsigned int buffer_size, const char* file);
    virtual unsigned int GetConfigurationInt(const char* section, const char* cle, unsigned int default_value, const char* file);
 
 protected:
